Add graddle support
