+++ src/main/java/duke/handlers/MarkAsDoneCommand.java
@@ -1,6 +1,6 @@
-package handlers;
+package duke.handlers;
 
-import models.TaskList;
+import duke.models.TaskList;
 
 public class MarkAsDoneCommand implements DukeCommand {
     public String run (TaskList taskList, String content) {
